refactor: use dirty instead of isEqual
