+++ src/components/pages/Users/UserPanel/UserPanel.test.tsx
@@ -11,8 +11,8 @@ import { Label as GroupsPanelFormLabel } from "components/GroupsPanelForm";
 import { Label as RolesPanelFormLabel } from "components/RolesPanelForm";
 import { renderComponent } from "test/utils";
 
+import { Label } from "./Fields/types";
 import UserPanel from "./UserPanel";
-import { Label } from "./types";
 
 const mockApiServer = setupServer(
   getGetGroupsMockHandler(),
