+++ src/components/pages/Users/UserPanel/index.ts
@@ -1,3 +1,3 @@
 export { default } from "./UserPanel";
 export type { Props as UserPanelProps } from "./types";
-export { Label as UserPanelLabel } from "./types";
+export { Label as UserPanelLabel } from "./Fields/types";
